flytte alerts i egen fil
